requirements.txt
asgiref==3.6.0
black==23.1.0
click==8.1.3
colorama==0.4.6
Django==4.1.7
django-cors-headers==3.14.0
django-environ==0.9.0
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
psycopg2-binary==2.9.3
python-http-client==3.3.7
pytz==2022.7.1
sendgrid==6.9.7
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
tomli==2.0.1
tzdata==2022.7
